refactor new code to vote_client
