+++ core/src/lib.rs
@@ -42,6 +42,7 @@ pub mod tvu;
 pub mod unfrozen_gossip_verified_vote_hashes;
 pub mod validator;
 mod vortexor_receiver_adapter;
+pub mod vote_client;
 pub mod vote_simulator;
 pub mod voting_service;
 pub mod warm_quic_cache_service;
